refactor: file reorg into Game folder
