+++ src/Game/utils/getInitialBoardSquares.ts
