+++ src/Game/utils/getNewValidLocation.ts
