+++ src/context/CurrentLevel.tsx
@@ -1,11 +1,11 @@
 import { createContext, useState, type ReactNode } from 'react';
-import levels from '../levels.json' with { type: 'json' };
+import levels from '../Game/levels.json' with { type: 'json' };
 import { colors } from '../CONSTANTS';
 export const CurrentLevelContext = createContext<CurrentLevelContextType>(
   {} as CurrentLevelContextType
 );
 import { InitialPiece } from '../types/piece';
-import { calculateUnitSize, findLargestHeight } from '../utils/utilities';
+import { calculateUnitSize, findLargestHeight } from '../Game/utils/utilities';
 import Hotjar from '@hotjar/browser';
 
 export interface CurrentLevelContextType {
